refactor: :recycle: split out routes into their own files
