+++ src/routes/greet.rs
@@ -1,0 +1,6 @@
+use actix_web::{HttpRequest, Responder};
+
+pub async fn greet(req: HttpRequest) -> impl Responder {
+    let name = req.match_info().get("name").unwrap_or("World");
+    format!("Hello {}!", &name)
+}
