+++ src/routes/mod.rs
@@ -1,0 +1,7 @@
+mod greet;
+mod health_check;
+mod subscriptions;
+
+pub use greet::*;
+pub use health_check::*;
+pub use subscriptions::*;
